electrod: internal UnixDomainSocket RPC Server
