+++ spruned/daemon/electrod/electrod_reactor.py
@@ -1,14 +1,14 @@
 import asyncio
-from typing import Tuple
 from spruned.abstracts import HeadersRepository
 from spruned.daemon.electrod.electrod_interface import ElectrodInterface
 from spruned.daemon.electrod.headers_repository import HeadersSQLiteRepository
+from spruned.daemon.electrod.electrod_rpc_server import ElectrodRPCServer
 from spruned import settings
 from spruned.daemon import database
-import aiomas
 
 
 class ElectrodReactor:
+    # FIXME - Move here the headers sync [out of the interface]
     def __init__(self, repo: HeadersRepository, interface, rpc_server, loop=None):
         self.repo = repo
         self.interface = interface
@@ -19,25 +19,6 @@ class ElectrodReactor:
         self.rpc_server.set_interface(self.interface)
         self.loop.create_task(self.interface.start())
         self.loop.create_task(self.rpc_server.start())
-
-
-class ElectrodRPCServer:
-    router = aiomas.rpc.Service()
-
-    def __init__(self, endpoint: (str, Tuple)):
-        self.endpoint = endpoint
-        self.interface: ElectrodInterface = None
-
-    def set_interface(self, interface: ElectrodInterface):
-        assert not self.interface, "RPC Server already initialized"
-        self.interface = interface
-
-    async def start(self):
-        return await aiomas.rpc.start_server(self.endpoint, self)
-
-    @router.expose
-    async def getrawtransaction(self, txid: str, verbose=False):
-        return await self.interface.getrawtransaction(txid)
 
 
 def build_electrod() -> ElectrodReactor:
